refactor: update imports for Page and ErrorPage components to new directory structure
