+++ src/app/screens/account/settings/About.tsx
@@ -2,8 +2,8 @@ import { useRef } from "react";
 import { useTranslation } from "react-i18next";
 import { Linking, ScrollView, Text, View } from "react-native";
 import { Button } from "@/components/common/Button";
-import { Page } from "@/components/common/Page";
 import { AnimatedLogo } from "@/components/custom/AnimatedLogo";
+import { Page } from "@/components/page/Page";
 import { useTheme } from "@/contexts/ThemeContext";
 
 export const About = () => {
