+++ src/app/screens/account/settings/ChangePassword.tsx
@@ -6,8 +6,8 @@ import { Keyboard } from "react-native";
 import { z } from "zod";
 import { Button } from "@/components/common/Button";
 import Input from "@/components/common/Input";
-import { Page } from "@/components/common/Page";
 import { useToast } from "@/components/common/Toast";
+import { Page } from "@/components/page/Page";
 import type { PasswordChange } from "@/dto";
 import { useChangePassword } from "@/hooks/account/useChangePassword";
 import { useUser } from "@/hooks/account/useUser";
