+++ src/app/screens/account/settings/Notifications.tsx
@@ -2,8 +2,8 @@ import { useQueryClient } from "@tanstack/react-query";
 import { useTranslation } from "react-i18next";
 import { Text, View } from "react-native";
 import Divider from "@/components/common/Divider";
-import { Page } from "@/components/common/Page";
 import { Switch } from "@/components/common/Switch";
+import { Page } from "@/components/page/Page";
 import { QUERY_KEYS } from "@/constants";
 import { useTheme } from "@/contexts/ThemeContext";
 import useNotification from "@/hooks/account/useNotification";
