+++ src/app/screens/account/settings/Statistics.tsx
@@ -3,7 +3,7 @@ import { useCallback, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Text, View } from "react-native";
 import { Button } from "@/components/common/Button";
-import { Page } from "@/components/common/Page";
+import { Page } from "@/components/page/Page";
 import { useTheme } from "@/contexts/ThemeContext";
 import { formatDate } from "@/utils";
