+++ src/app/screens/auth/ResetPassword.tsx
@@ -13,8 +13,8 @@ import {
 import { z } from "zod";
 import { Button } from "@/components/common/Button";
 import Input from "@/components/common/Input";
-import { Page } from "@/components/common/Page";
 import { useToast } from "@/components/common/Toast";
+import { Page } from "@/components/page/Page";
 import { useTheme } from "@/contexts/ThemeContext";
 import useAuth from "@/hooks/account/useAuth";
 import type { AuthStackParamList } from "@/services/storage/types";
