+++ src/app/screens/auth/Signup.tsx
@@ -13,8 +13,8 @@ import { z } from "zod";
 import { VerificationCodeModal } from "@/components/auth/VerificationCode";
 import { Button } from "@/components/common/Button";
 import Input from "@/components/common/Input";
-import { Page } from "@/components/common/Page";
 import { useToast } from "@/components/common/Toast";
+import { Page } from "@/components/page/Page";
 import { useTheme } from "@/contexts/ThemeContext";
 import useAuth from "@/hooks/account/useAuth";
 import i18n from "@/i18n";
