+++ src/app/screens/services/Services.tsx
@@ -6,9 +6,9 @@ import { View } from "react-native";
 import Animated from "react-native-reanimated";
 import { IconButton } from "@/components/common/Button";
 import Card from "@/components/common/Card";
-import { Page } from "@/components/common/Page";
-import { PreferenceCustomizationButton } from "@/components/common/PreferenceCustomizationModal";
-import { Empty } from "@/components/custom/Empty";
+import { PreferenceCustomizationButton } from "@/components/custom/widget/PreferenceCustomizationModal";
+import { Empty } from "@/components/page/Empty";
+import { Page } from "@/components/page/Page";
 import { useTheme } from "@/contexts/ThemeContext";
 import { useAnimatedHeader } from "@/hooks/useAnimatedHeader";
 import { useServicePreferences } from "@/hooks/usePreferences";
