+++ src/app/screens/services/homework/components/ErrorState/ErrorState.tsx
@@ -1,7 +1,7 @@
 import { t } from "i18next";
 import { Text, View } from "react-native";
-import { Page } from "@/components/common/Page";
 import { AboutModal } from "@/components/custom/AboutModal";
+import { Page } from "@/components/page/Page";
 
 export const ErrorState = ({
   error,
