+++ src/app/screens/services/restaurant/components/LoadingState/LoadingState.tsx
@@ -1,9 +1,9 @@
 import { useMemo } from "react";
 import { useTranslation } from "react-i18next";
 import { Text, View } from "react-native";
-import { Page } from "@/components/common/Page";
 import { AboutModal } from "@/components/custom/AboutModal";
 import RestaurantCard from "@/components/custom/card/RestaurantCard";
+import { Page } from "@/components/page/Page";
 import { getOpeningHoursData } from "@/utils";
 
 export const LoadingState = () => {
