+++ src/app/screens/services/traq/Traq.tsx
@@ -1,7 +1,7 @@
 import { useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { Page } from "@/components/common/Page";
-import { ErrorPage } from "@/components/custom/ErrorPage";
+import { ErrorPage } from "@/components/page/ErrorPage";
+import { Page } from "@/components/page/Page";
 import { useTraq } from "@/hooks/useTraq";
 import {
   AboutSection,
