+++ src/app/screens/services/traq/components/LoadingState/LoadingState.tsx
@@ -1,8 +1,8 @@
 import { useTranslation } from "react-i18next";
 import { ScrollView, View } from "react-native";
 import { BadgeLoading } from "@/components/common/Badge";
-import { Page } from "@/components/common/Page";
 import { TraqCardLoading } from "@/components/custom/card/TraqCard";
+import { Page } from "@/components/page/Page";
 import { TextSkeleton } from "@/components/Skeleton";
 import { AboutSection } from "../AboutSection";
