+++ src/app/screens/services/washing-machines/components/LoadingState/LoadingState.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next";
 import { Text, View } from "react-native";
-import { Page } from "@/components/common/Page";
 import { WashingMachineCardSkeleton } from "@/components/custom/card/WashingMachineCard";
+import { Page } from "@/components/page/Page";
 import { useTheme } from "@/contexts/ThemeContext";
 import { AboutSection } from "../AboutSection";
